fs_python/requirements.txt
Flask==2.1.3
google-cloud-storage
python-dotenv
gunicorn